feat: show workspace picker content in left nav
