+++ changelogs/fragments/7716.yml
@@ -1,0 +1,2 @@
+feat:
+- Display workspace picker content when outside workspace ([#7716](https://github.com/opensearch-project/OpenSearch-Dashboards/pull/7716))
